abstract 404 page layout
