+++ src/pages/404.js
@@ -1,28 +1,7 @@
 import React from 'react';
-import { Grid } from '../components/grid';
 import { StaticQuery, graphql } from 'gatsby';
-import Helmet from 'react-helmet';
-import ReactMarkdown from 'react-markdown';
 
-import Layout from '../components/layout';
-
-import {
-  StyledBlueBackground,
-  StyledRow,
-  TextCol,
-  IllustrationCol,
-  TitleHeadline,
-  CopyText,
-  LinksTitle,
-  LinkParagraph,
-  NotFoundPageLink,
-} from '../components/404';
-
-import { LogoStyleContext } from '../context/PageContext';
-
-import illustration from '../images/404-illustration.svg';
-
-import { colors } from '../utils/theme';
+import { LinksTitle, LinkParagraph, NotFoundPageLink } from '../components/404';
 
 // Page
 
@@ -55,58 +34,23 @@ const NotFoundPage = () => (
         linksTitle,
       } = content;
       return (
-        <LogoStyleContext.Provider
-          value={{
-            fillColorInitial: colors.white,
-            textColor: colors.blueBg,
-          }}
+        <NotFoundPage
+          siteTitle={site.siteMetadata.siteTitle}
+          footerId={id}
+          titleHeadline={title}
+          copy={copy}
         >
-          <Layout is404={true} bgColor="blueBg" footerContactUsId={id}>
-            <Helmet
-              title={`${site.siteMetadata.siteTitle} - Not Found`}
-              meta={[
-                {
-                  name: 'description',
-                  content: 'YLD - Engineering - Digital, NodeJS, React, AWS',
-                },
-              ]}
-            >
-              <html lang="en" />
-            </Helmet>
-            <StyledBlueBackground>
-              <Grid>
-                <StyledRow>
-                  <TextCol>
-                    <TitleHeadline>{title}</TitleHeadline>
-                    <ReactMarkdown
-                      renderers={{
-                        paragraph: CopyText,
-                      }}
-                      source={copy}
-                    />
-                    <LinksTitle>{linksTitle}</LinksTitle>
-                    <LinkParagraph>
-                      <NotFoundPageLink to="/">Home</NotFoundPageLink>
-                    </LinkParagraph>
-                    <LinkParagraph>
-                      <NotFoundPageLink to="/our-work/">
-                        Our work
-                      </NotFoundPageLink>
-                    </LinkParagraph>
-                    <LinkParagraph>
-                      <NotFoundPageLink to="/contact/">
-                        Contact
-                      </NotFoundPageLink>
-                    </LinkParagraph>
-                  </TextCol>
-                  <IllustrationCol>
-                    <img src={illustration} alt="" />
-                  </IllustrationCol>
-                </StyledRow>
-              </Grid>
-            </StyledBlueBackground>
-          </Layout>
-        </LogoStyleContext.Provider>
+          <LinksTitle>{linksTitle}</LinksTitle>
+          <LinkParagraph>
+            <NotFoundPageLink to="/">Home</NotFoundPageLink>
+          </LinkParagraph>
+          <LinkParagraph>
+            <NotFoundPageLink to="/our-work/">Our work</NotFoundPageLink>
+          </LinkParagraph>
+          <LinkParagraph>
+            <NotFoundPageLink to="/contact/">Contact</NotFoundPageLink>
+          </LinkParagraph>
+        </NotFoundPage>
       );
     }}
   />
